Add env and cluster app
