--- packages/compile-service/.env.example
@@ -1,4 +1,0 @@
-NODE_ENV="development"
-MONGODB_URL="mongodb://localhost:27017/zk?directConnection=true"
-REDIS_URL="redis://localhost:6379"
-PROOF_MONGODB_URL="mongodb://"
