--- packages/compile-service/package.json
@@ -1,31 +1,0 @@
-{
-  "name": "@zkdb/compile-service",
-  "version": "1.0.0",
-  "main": "index.js",
-  "license": "MIT",
-  "type": "module",
-  "dependencies": {
-    "@orochi-network/vault": "^1.4.15",
-    "@zkdb/smart-contract": "^1.1.6",
-    "@zkdb/storage": "^1.1.0",
-    "bson": "6.10.1",
-    "bullmq": "5.34.2",
-    "redis": "4.7.0"
-  },
-  "scripts": {
-    "dev": "tsx ./src/index.ts",
-    "start": "node ./build/src/index.js",
-    "build": "rm -rf ./build && rollup -c",
-    "start:prod": "yarn build && yarn start"
-  },
-  "peerDependencies": {
-    "o1js": "^1.9.*"
-  },
-  "devDependencies": {
-    "@rollup/plugin-alias": "^5.1.1",
-    "@rollup/plugin-commonjs": "^28.0.1",
-    "@rollup/plugin-typescript": "^12.1.1",
-    "@types/node": "^22.7.7",
-    "rollup": "^4.24.0"
-  }
-}
