--- packages/compile-service/src/helper/index.ts
@@ -1,1 +1,0 @@
-export * from "./logger";
