--- packages/compile-service/src/helper/logger.ts
@@ -1,7 +1,0 @@
-import { LoggerLoader } from "@orochi-network/framework";
-import { config } from "./config";
-export const logger = new LoggerLoader(
-  "zkDatabase",
-  config.LOG_LEVEL,
-  "string"
-);
