--- packages/compile-service/src/service/index.ts
@@ -1,1 +1,0 @@
-export * from "./zk-compile";
