--- packages/compile-service/tsconfig.json
@@ -1,27 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "compilerOptions": {
-    "target": "ESNext",
-    "module": "ESNext",
-    "outDir": "./build",
-    "moduleResolution": "node",
-    "esModuleInterop": true,
-    "experimentalDecorators": true,
-    "emitDecoratorMetadata": true,
-    "strictPropertyInitialization": false,
-    "noFallthroughCasesInSwitch": true,
-    "types": ["node"],
-    "composite": true,
-    "declarationMap": false,
-    "preserveSymlinks": true,
-    "rootDir": ".",
-    "baseUrl": "./",
-    "paths": {
-      "@helper": ["./src/helper"],
-      "@service": ["./src/service"]
-    },
-  },
-  "references": [{ "path": "../smart-contract" }, { "path": "../storage" }],
-  "include": ["**/*.ts"],
-  "exclude": ["build", "node_modules"]
-}
