+++ packages/mina-service/.env.example
@@ -1,4 +1,5 @@
 MONGODB_URL="mongodb://"
 NODE_ENV="development"
-BROKER_SERVICE='http://0.0.0.0:31337'
 PROOF_MONGODB_URL="mongodb://"
+REDIS_URL="redis://localhost:6379"
+SERVICE_SECRET="base64string"
