+++ packages/mina-service/package.json
@@ -57,7 +57,8 @@
     "typescript": "^5.6.2"
   },
   "dependencies": {
-    "@orochi-network/framework": "^2.3.19",
+    "@orochi-network/framework": "^2.3.22",
+    "@orochi-network/vault": "^1.4.15",
     "@zkdb/smart-contract": "^1.1.0",
     "@zkdb/storage": "^1.1.0",
     "bullmq": "^5.34.3",
