+++ packages/mina-service/src/service/index.ts
@@ -1,1 +1,2 @@
 export * from './task-service';
+export * from './compile-service';
